Move the v3 specific observer into a new v3 package
