+++ pkg/v3/observer.go
@@ -1,29 +1,33 @@
-package ocr2keepers
+package v3
 
-import "context"
+import (
+	"context"
+
+	ocr2keepers "github.com/smartcontractkit/ocr2keepers/pkg"
+)
 
 type Tick interface {
 	// GetUpkeeps provides an array of upkeeps scoped to the individual tick
-	GetUpkeeps(ctx context.Context) ([]UpkeepPayload, error)
+	GetUpkeeps(ctx context.Context) ([]ocr2keepers.UpkeepPayload, error)
 }
 
 // Preprocessor is the general interface for middleware used to filter, add, or modify upkeep
 // payloads before checking their eligibility status
 type Preprocessor interface {
 	// PreProcess takes a slice of payloads and returns a new slice
-	PreProcess(context.Context, []UpkeepPayload) ([]UpkeepPayload, error)
+	PreProcess(context.Context, []ocr2keepers.UpkeepPayload) ([]ocr2keepers.UpkeepPayload, error)
 }
 
 // Postprocessor is the general interface for a processing function after checking eligibility status
 type Postprocessor interface {
 	// PostProcess takes a slice of results where eligibility status is known
-	PostProcess(context.Context, []CheckResult) error
+	PostProcess(context.Context, []ocr2keepers.CheckResult) error
 }
 
 // Runner2 is the interface for an object that should determine eligibility state
 type Runner2 interface {
 	// CheckUpkeeps has an input of upkeeps with unknown state and an output of upkeeps with known state
-	CheckUpkeeps(context.Context, []UpkeepPayload) ([]CheckResult, error)
+	CheckUpkeeps(context.Context, []ocr2keepers.UpkeepPayload) ([]ocr2keepers.CheckResult, error)
 }
 
 type Observer struct {
@@ -49,7 +53,7 @@ func (o *Observer) Process(ctx context.Context, tick Tick) error {
 		return err
 	}
 
-	var upkeepPayloads []UpkeepPayload
+	var upkeepPayloads []ocr2keepers.UpkeepPayload
 	// Run pre-processors
 	for _, preprocessor := range o.Preprocessors {
 		upkeepPayloads, err = preprocessor.PreProcess(ctx, upkeeps)
